"updated the imports elsewhere in order to make the refactoring clean"
