+++ src/app/app.module.ts
@@ -24,7 +24,7 @@ import { TranslateModule, TranslateLoader, TranslateService } from '@ngx-transla
 import { TranslationService } from './core/services/translation.service';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { ElapsedTimePipe } from './shared/pipes/elapsed-time.pipe';
-import { RegisterComponent } from './core/pages/register/register.component'; 
+import { RegisterComponent } from './pages/register/register.component'; 
 
 
 // Define a function that invoke TranslationService
